Only delete target file if exists (#748)
Corner case where a download failure would cause darc to crash

+++ src/Microsoft.DotNet.Darc/src/Darc/Operations/GatherDropOperation.cs
@@ -1238,7 +1238,10 @@ namespace Microsoft.DotNet.Darc.Operations
             {
                 // Ensure we delete the file in this case, otherwise an attempt to download
                 // from a separate location will fail.
-                File.Delete(targetFile);
+                if (File.Exists(targetFile))
+                {
+                    File.Delete(targetFile);
+                }
                 errors.Add($"Failed to download {sourceUri}: {e.Message}");
             }
             finally
